Refactored codebase for improved maintainability
  - Removed dead code from errors.rs (5 unused error variants, 2
    extension traits)
  - Removed unused fields and methods from app_state.rs (url field, 7
    getter methods)
  - Extracted display utilities to src/utils/display.rs
    (truncate_url_for_display)
  - Split tui.rs into submodules (mod.rs, render.rs, input.rs)
  - Simplified settings_menu.rs with helper functions and data-driven
    option validation
  - Added should_abort() and log_msg() helpers in worker.rs to reduce
    boilerplate
  - Updated progress_parser.rs to remove unused url parameter
