+++ src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod dependencies;
+pub mod display;
 pub mod file;
 pub mod settings;
